BREAKING CHANGE: Rename kinesis exporter to awskinesis
Addresses #2164
