+++ exporter/awskinesisexporter/README.md
